Ashanti's BSearch Code
If list is null, it skips to the end and prints -1. Otherwise, the bsearch is standard. Originally, the while loop would continue forever if x was not in the list, but I made it so that the while loop can't iterate more than [a] times, a being the length of the list.